CTC-17 Extract scrapping of git url
